requirements.txt
asgiref==3.8.1
Django==5.1.2
django-autoslug==1.9.9
pillow==11.0.0
sqlparse==0.5.1
tzdata==2024.2
